+++ grunt.js
@@ -2,11 +2,12 @@
 module.exports = function(grunt) {
 
   grunt.loadNpmTasks('grunt-rigger');
+  grunt.loadNpmTasks('grunt-jasmine-runner');
 
   // Project configuration.
   grunt.initConfig({
     meta: {
-      version: '0.9.11',
+      version: '0.10.2',
       banner: '// Backbone.Marionette, v<%= meta.version %>\n' +
         '// Copyright (c)<%= grunt.template.today("yyyy") %> Derick Bailey, Muted Solutions, LLC.\n' + 
         '// Distributed under MIT license\n' + 
@@ -14,7 +15,7 @@ module.exports = function(grunt) {
     },
 
     lint: {
-      files: ['src/backbone.marionette.js']
+      files: ['src/backbone.marionette.*.js']
     },
 
     rig: {
@@ -25,10 +26,6 @@ module.exports = function(grunt) {
       amd: {
         src: ['<banner:meta.banner>', 'src/amd.js'],
         dest: 'lib/amd/backbone.marionette.js'
-      },
-      async: {
-        src: ['<banner:meta.banner>', 'src/async/async.js'],
-        dest: 'lib/backbone.marionette.async.js'
       }
     },
 
@@ -40,11 +37,40 @@ module.exports = function(grunt) {
       amd: {
         src: ['<banner:meta.banner>', '<config:rig.amd.dest>'],
         dest: 'lib/amd/backbone.marionette.min.js'
-      },
-      async: {
-        src: ['<banner:meta.banner>', '<config:rig.async.dest>'],
-        dest: 'lib/backbone.marionette.async.min.js'
       }
+    },
+
+    jasmine : {
+      src : [
+        'public/javascripts/jquery.js',
+        'public/javascripts/json2.js',
+        'public/javascripts/underscore.js',
+        'public/javascripts/backbone.js',
+        'src/backbone.marionette.js',
+        'src/async/async.js',
+        'spec/javascripts/support/marionette.support.js',
+        'src/backbone.marionette.eventbinder.js',
+        'src/backbone.marionette.view.js',
+        'src/backbone.marionette.itemview.js',
+        'src/backbone.marionette.collectionview.js',
+        'src/backbone.marionette.compositeview.js',
+        'src/backbone.marionette.region.js',
+        'src/backbone.marionette.layout.js',
+        'src/backbone.marionette.application.js',
+        'src/backbone.marionette.approuter.js',
+        'src/backbone.marionette.module.js',
+        'src/backbone.marionette.templatecache.js',
+        'src/backbone.marionette.renderer.js',
+        'src/backbone.marionette.callbacks.js',
+        'src/backbone.marionette.eventaggregator.js',
+        'src/backbone.marionette.helpers.js'
+      ],
+      helpers : 'spec/javascripts/helpers/*.js',
+      specs : 'spec/javascripts/**/*.spec.js'
+    },
+
+    'jasmine-server' : {
+      browser : false
     },
 
     jshint: {
@@ -64,7 +90,10 @@ module.exports = function(grunt) {
       globals: {
         jQuery: true,
         Backbone: true,
-        _: true
+        _: true,
+        Marionette: true,
+        $: true,
+        slice: true
       }
     },
     uglify: {}