+++ spec/javascripts/layout.spec.js
@@ -1,9 +1,15 @@
 describe("layout", function(){
+
   var LayoutManager = Backbone.Marionette.Layout.extend({
     template: "#layout-manager-template",
     regions: {
       regionOne: "#regionOne",
       regionTwo: "#regionTwo"
+    },
+    initialize: function() {
+      if (this.model) {
+        this.bindTo(this.model, 'change', this.render);
+      }
     }
   });
 
@@ -52,7 +58,6 @@ describe("layout", function(){
       expect(layoutManager).toHaveOwnProperty("regionOne");
       expect(layoutManager).toHaveOwnProperty("regionTwo");
     });
-
 
   });
 
@@ -167,7 +172,9 @@ describe("layout", function(){
     beforeEach(function(){
       loadFixtures("layoutManagerTemplate.html");
 
-      layout = new LayoutManager();
+      layout = new LayoutManager({
+        model: new Backbone.Model()
+      });
       layout.render();
       region = layout.regionOne;
 
@@ -194,6 +201,21 @@ describe("layout", function(){
       var regionEl = layout.$("#regionOne");
       expect(region.$el[0]).toBe(regionEl[0]);
     });
+
+    describe("and the view's `render` function is bound to an event in the `initialize` function", function(){
+      beforeEach(function(){
+        layout.onRender = function() {
+          this.regionOne.show(view);
+        };
+
+        layout.model.trigger('change');
+      });
+
+      it("should re-bind the regions correctly", function(){
+        expect(layout.$("#regionOne")).not.toBeEmpty();
+      });
+    });
+
   });
 
   describe("when re-rendering a closed layout", function(){